[#435] REFACTOR: Authentication filter에서 발생한 에러 처리 분리
