+++ backend/src/main/java/com/team5/secondhand/global/auth/AuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.team5.secondhand.global.jwt.config;
+package com.team5.secondhand.global.auth;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.team5.secondhand.application.member.dto.response.MemberDetails;
@@ -7,22 +7,26 @@ import com.team5.secondhand.global.jwt.util.JwtUtils;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import javax.servlet.*;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
 @Component
+@Order(2)
 @RequiredArgsConstructor
 public class AuthenticationFilter extends OncePerRequestFilter {
 
     private final JwtProperties jwtProperties;
     private final ObjectMapper objectMapper;
     private final JwtUtils jwtUtils;
+    private final AuthenticationContext context;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws IOException, ServletException {
@@ -34,7 +38,7 @@ public class AuthenticationFilter extends OncePerRequestFilter {
             Claims claim = jwtUtils.getClaim(token);
             loginMember = objectMapper.readValue((String) claim.get(jwtProperties.getClaimKey()), MemberDetails.class);
         }
-        request.setAttribute("loginMember", loginMember);
+        context.storeLoginMember(loginMember);
 
         chain.doFilter(request, response);
     }
