+++ backend/src/main/java/com/team5/secondhand/global/auth/ExceptionHandleFilter.java
@@ -1,0 +1,29 @@
+package com.team5.secondhand.global.auth;
+
+import io.jsonwebtoken.JwtException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.annotation.Order;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+@Component
+@Order(1)
+public class ExceptionHandleFilter extends OncePerRequestFilter {
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        try {
+            filterChain.doFilter(request, response);
+        } catch (JwtException e) {
+            log.error("🔐토큰이 잘못됐어요!!");
+            response.sendError(HttpServletResponse.SC_UNAUTHORIZED, e.getMessage());
+        }
+    }
+}
