+++ backend/src/main/java/com/team5/secondhand/global/exception/GlobalExceptionHandler.java
@@ -2,7 +2,6 @@ package com.team5.secondhand.global.exception;
 
 import com.team5.secondhand.global.dto.ErrorResponse;
 import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.JwtException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
@@ -22,14 +21,9 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(AuthenticationException.class)
     @ResponseStatus(HttpStatus.UNAUTHORIZED)
     public ErrorResponse handleAuthenticationException(AuthenticationException e) {
-          return ErrorResponse.occur(e);
-    }
-
-    @ExceptionHandler(JwtException.class)
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    public ErrorResponse handleJwtException(JwtException e) {
         return ErrorResponse.occur(e);
     }
+
     @ExceptionHandler(ExpiredJwtException.class)
     @ResponseStatus(HttpStatus.UNAUTHORIZED)
     public ErrorResponse handleJExpiredJwtException(ExpiredJwtException e) {
